requirements.txt
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
psycopg2==2.8.6
pycparser==2.20
python-dotenv==0.17.1
six==1.16.0
